==> scripts_v2/10_data_to_sql.ipynb <==
import pandas as pd
import json
import sqlite3
import numpy as np
from datetime import datetime

# Load in the data

recipes = pd.read_csv("../data/RecSys/raw-data_recipe.csv")

interactions = pd.read_csv("../data/RecSys/raw-data_interaction.csv")

# Format Recipes

## Rename Columns

I rename some of the columns for ease. I won't be using the reviews here so dropping that.

recipes = recipes.rename(columns={'recipe_id':'id', 'recipe_name':'name', 'aver_rate':'ave_rating', 'cooking_directions':'directions', 'nutritions':'nutrients'})
recipes = recipes.drop('reviews', axis=1)
recipes.head()

## Cooking Directions

Format the directions. See the effect.

directions = [ x.replace("{'directions': u", "").replace("'}", '') for x in recipes.directions ]
recipes['directions'] = directions

recipes.head()

## Ingredients

Format the ingredients. It has ^ to separate the individual ingredients. Want to use ',' instead. There are 2 current instances of commas so just replace those with a ' '.

So first find and replace the existing commas.

any_commas = np.array([ x.find(',') for x in recipes.ingredients ])
np.sum(any_commas > (-1))

# Where are those commas
for x in recipes.ingredients[any_commas > (-1)]:
    print(x.split('^'))

for ind in np.nonzero(any_commas > (-1))[0]:
    recipes.loc[ind,"ingredients"] = recipes.loc[ind,"ingredients"].replace(",", " ")

Replace out all ^ with ,

recipes['ingredients'] = [ x.replace('^', ', ') for x in recipes.ingredients ]

recipes.head()

## Format Interactions

This is mostly dealing with the dateLastModified column.

interactions.head()

# Strip the \n at the end of dateLastModified
interactions['dateLastModified'] = [ x.strip('\n') for x in interactions.dateLastModified ]

* The date last modified includes time-scales faster than seconds. 
* Those microseconds aren't needed + are missing on many of the rows. 
* So we will restrict the date part to only a length of 19

# We can the length of the dateLastModified
lens = pd.Series([ len(x) for x in interactions.dateLastModified ])
lens.value_counts()

# We want to limit the length to 19
interactions['dateLastModified'] = [ x[:19] for x in interactions.dateLastModified ]

# Convert datetime into integer
interactions['dateLastModified'] = [ datetime.fromisoformat(x).strftime('%s') for x in interactions.dateLastModified ]

interactions.head()

## Format Nutrients

## Compile Nutrients Separately

def convert_nutrients_row(recipe_id, nutrients):
    """
    Takes in a json string of the nutrients for a given recipe. 
    Returns a dataframe with rows being each nutrient + columns giving the information.
    """
    # Clean
    nutrients_clean = nutrients.replace("'", '"').replace("u","").replace("True","true").replace("False","false").replace("Tre","true")
    # To Dist
    d = json.loads(nutrients_clean)
    # To DataFrame
    df = pd.DataFrame(list(d.values()))
    # Rename
    df = df.rename(columns={'hasCompleteData': 'complete', 'amont': 'value', 'percentDailyVale': 'percent_daily_value', 
                       'displayVale': 'display_value', 'nit': 'value_unit'})
    # Add the recipe
    df.insert(0, 'recipe_id', recipe_id)
    return df

# Check if one recipe works
convert_nutrients_row(recipes.id[0], recipes.nutrients[0]).head()

lst_nutrients = []
bad_inds = [] # Nutrient information isn't available for these ids
for index,row in recipes.iterrows():
    #print(index)
    try:
        df_nutrients = convert_nutrients_row(row['id'], row['nutrients'])
        lst_nutrients.append(df_nutrients)
    except:
        bad_inds.append(index)

# Some of the recipes are missing nutrient information as shown here
recipes.iloc[bad_inds[10],:]['nutrients']

# Prop of recipes with nutrient info
np.round(len(lst_nutrients)/(len(lst_nutrients) + len(bad_inds)), 2)

# Nutrients
nutrients = pd.concat(lst_nutrients)
nutrients.head()

## Format some nutrient columns

# Fix the percent value
nutrients.loc[nutrients.percent_daily_value == '< 1', 'percent_daily_value'] = '0'
nutrients.loc[nutrients.percent_daily_value == '-', 'percent_daily_value'] = np.NaN
nutrients = nutrients.astype({'percent_daily_value': 'float32'})

# I want to add a nutrient id => labels
labels, levels = pd.factorize(nutrients.name)
labels += 1
## like R's factor function
print(labels)
print(levels)
## Add the labels as a new column
nutrients.insert(0, 'nutrient_id', labels)

nutrients.head()

# Remove the nutrients from the recipes
recipes = recipes.drop('nutrients', axis=1)
recipes.head()

# Send to SQL

conn = sqlite3.connect('../data/recsys.db')
c = conn.cursor()

c.execute('DROP TABLE IF EXISTS recipes;')
c.execute('DROP TABLE IF EXISTS interactions;')
c.execute('DROP TABLE IF EXISTS nutrients;')
conn.commit()

# Create table for recipes
rec_sql = '''CREATE TABLE recipes(
                id INTEGER PRIMARY KEY,
                name TEXT NOT NULL,
                ave_rating REAL,
                review_nums INTEGER,
                image_url TEXT,
                ingredients TEXT,
                directions TEXT)'''

int_sql = '''CREATE TABLE interactions(
                user_id INTEGER NOT NULL,
                recipe_id INTEGER NOT NULL, 
                rating INTEGER NOT NULL, 
                date_mod INTEGER NOT NULL,
                PRIMARY KEY (user_id, recipe_id),
                FOREIGN KEY (recipe_id) REFERENCES recipes (id) 
                    ON DELETE CASCADE ON UPDATE NO ACTION)'''

nutr_sql = '''CREATE TABLE nutrients(
                nutrient_id INTEGER NOT NULL, 
                recipe_id INTEGER NOT NULL, 
                complete INTEGER DEFAULT 0, 
                name TEXT NOT NULL, 
                value REAL NOT NULL, 
                percent_daily_value INTEGER NOT NULL, 
                display_value TEXT NOT NULL, 
                value_unit TEXT NOT NULL, 
                PRIMARY KEY (nutrient_id, recipe_id), 
                FOREIGN KEY (recipe_id) REFERENCES recipes (id)
                    ON DELETE CASCADE ON UPDATE NO ACTION)'''

c.execute(rec_sql)
c.execute(int_sql)
c.execute(nutr_sql)

conn.commit()

recipes.to_sql('recipes', conn, if_exists='replace', index=False)

interactions.to_sql('interactions', conn, if_exists='replace', index=False)

nutrients.to_sql('nutrients', conn, if_exists='replace', index=False)
