==> scripts_v2/14_nutrient_data.ipynb <==
Relevant searches to extract data structures that can be of use down the line.

import pandas as pd
import json
import sqlite3
import numpy as np
from datetime import datetime

conn = sqlite3.connect('../data/recsys.db')
c = conn.cursor()

# Nutrient Data

1. Load the nutrient data and pivot (long => wide).
2. Then join this with the original recipe data. This can be saved out and sent.

df = pd.read_sql_query('''SELECT * FROM nutrients;''', conn)
df.head(30)

df_val = df.pivot(index='recipe_id', columns='name', values='value')
df_val.head()

# Only get those recipes with some sort-of nutrient info
df_recipes = pd.read_sql_query('''SELECT * FROM recipes WHERE id in (SELECT DISTINCT(recipe_id) FROM nutrients);''', conn)
df_recipes.head()

pd.merge(df, df_recipes, )

df2 = pd.merge(df_recipes, df_val, left_on='id', right_on='recipe_id')
df2.head()

df2.shape

df2.to_csv('../data/14_recipes_with_nutrients.csv')
