==> scripts_v2/16_interaction_data.ipynb <==
import pandas as pd
import json
import sqlite3
import numpy as np
from datetime import datetime

conn = sqlite3.connect('../data/recsys.db')
c = conn.cursor()

## Interaction Data

We want to separate the user data into those that exist before 2013 (training) and those that exist after 2013 (test). We remove users with less than 3 ratings in each of those time bins. In addition, we want to use this data to simulate paired choice behavior (that is given 2 recipes, which item will a person choose) and so we need some amount of variation in the data. That is each user needs more than 1 unique rating value. So we also exclude any user where all the ratings are the same (0 variance).

c.execute('''DROP TABLE IF EXISTS pre_interactions;''')
c.execute('''DROP TABLE IF EXISTS post_interactions;''')
c.execute('''/* Pre-2013 - User Ids - Variance of Ratings > 0 */
CREATE TEMPORARY TABLE pre_interactions AS
SELECT user_id, (AVG(rating*rating) - AVG(rating)*AVG(rating)) AS var_rating, COUNT(*) AS n_ratings
FROM interactions
WHERE (dateLastModified - strftime('%s','2013-01-01')) < 0
GROUP BY user_id
HAVING var_rating > 0 AND n_ratings > 2;''')
c.execute('''/* Post-2013 - User Ids - Variance of Ratings > 0 */
CREATE TEMPORARY TABLE post_interactions AS
SELECT user_id, (AVG(rating*rating) - AVG(rating)*AVG(rating)) AS var_rating, COUNT(*) AS n_ratings
FROM interactions
WHERE (dateLastModified - strftime('%s','2013-01-01')) >= 0
GROUP BY user_id
HAVING var_rating > 0 AND n_ratings > 2;''')
conn.commit()

df = pd.read_sql_query('''/* Ok so now I want those user_ids that exist both pre-2013 and post-2013 */
/* And then I want to get all the ratings for those user-ids */
SELECT user_id, recipe_id, rating, dateLastModified AS date_int, datetime(dateLastModified, 'unixepoch') AS date, 
	(CASE WHEN (dateLastModified - strftime('%s','2013-01-01')) < 0 THEN 1 ELSE 0 END) AS training
FROM interactions
	INNER JOIN pre_interactions AS prei
		USING (user_id)
	INNER JOIN post_interactions AS posti
		USING (user_id)
ORDER BY user_id, date ASC;''', conn)
df.head()

df.groupby('user_id').agg('count')

df.groupby('training').agg(['count', 'nunique'])

404762/(404762 + 187660) # percent of training

df.to_csv('../data/16_interaction_data.csv')
